VTCCA-3824 bugs fixed
