+++ app/actions/requests/BasicAuthenticatedRequest.scala
@@ -20,13 +20,6 @@ import models.{Accounts, DetailedIndividualAccount, GroupAccount}
 import play.api.mvc.{Request, WrappedRequest}
 import uk.gov.hmrc.auth.core.AffinityGroup
 
-abstract class AuthenticatedRequest[A](request: Request[A]) extends WrappedRequest[A](request) {
-  def organisationAccount: GroupAccount
-  def individualAccount: DetailedIndividualAccount
-  def organisationId: Long = organisationAccount.id
-  def personId: Long = individualAccount.individualId
-}
-
 //need to decouple to authentication away from the group account and individual account.
 case class BasicAuthenticatedRequestWithAffinityGroup[A](
       organisationAccount: GroupAccount,
